+++ cx-portal/src/components/pages/AppOverview/index.tsx
@@ -35,6 +35,8 @@ import './AppOverview.scss'
 import { useFetchProvidedAppsQuery } from 'features/apps/apiSlice'
 import { useDispatch } from 'react-redux'
 import debounce from 'lodash.debounce'
+import { OVERLAYS } from 'types/Constants'
+import { show } from 'features/control/overlay/actions'
 
 export default function AppOverview() {
   const { t } = useTranslation()
@@ -124,6 +126,12 @@ export default function AppOverview() {
     [debouncedSearch, items, group]
   )
 
+  const showOverlay = (item: any) => {
+    if (item.status === 'created') {
+      dispatch(show(OVERLAYS.APP_OVERVIEW_CONFIRM, item.name))
+    }
+  }
+
   return (
     <div className="appoverview-app">
       <PageHeader
@@ -150,6 +158,9 @@ export default function AppOverview() {
                 variant="minimal"
                 filledBackground={true}
                 imageSize={'small'}
+                onCardClick={(item: any) => {
+                  showOverlay(item)
+                }}
               />
             </div>
           </div>
@@ -194,6 +205,9 @@ export default function AppOverview() {
                 showAddNewCard={true}
                 newButtonText={t('content.appoverview.addbtn')}
                 onNewCardButton={function noRefCheck() {}}
+                onCardClick={(item: any) => {
+                  showOverlay(item)
+                }}
               />
             </div>
           )}